add: change external script

+++ src/cvt_file.py
@@ -16,29 +16,13 @@ def cvtScriptFile2Key(script: ScriptSection) -> FileKeyScriptSection:
     return output
 
 
-class Action:
-    src_hash: str
-    src_ext: str
-    dst_hash: str
-    dst_ext: str
-
-    def __init__(
-        self,
-        src_hash: str,
-        dst_hash: str,
-        src_ext: str,
-        dst_ext: str,
-    ) -> None:
-        self.src_ext = src_ext
-        self.src_hash = src_hash
-        self.dst_ext = dst_ext
-        self.dst_hash = dst_hash
-
-
 def convert_hash(
-    from_scritp: ScriptSection, to_script: FileKeyScriptSection, mod_dir: str
+    config: Config, from_scritp: ScriptSection, to_script: FileKeyScriptSection
 ):
-    actions: dict[str, Action] = {}
+    actions: HashedActions = {}
+    assert config.mod is not None
+    mod_dir = config.mod
+    print(mod_dir)
     assert path.isdir(mod_dir)
     for ext in listdir(mod_dir):
         dir = path.join(mod_dir, ext)
@@ -46,7 +30,7 @@ def convert_hash(
             continue
         for item in listdir(dir):
             fsplit = item.split(".")
-            if len(fsplit) != 2:
+            if len(fsplit) != 2 or fsplit[1] != "wem":
                 continue
 
             from_hash = fsplit[0]
@@ -56,21 +40,38 @@ def convert_hash(
             f_key = rmLangPath(from_scritp[from_hash]["fileName"])
             new_hash = to_script[f_key]["hash"]
 
-            actions[from_hash] = Action(
+            actions[new_hash] = Action(
                 src_ext=ext, dst_ext=ext, src_hash=from_hash, dst_hash=new_hash
             )
     return actions
 
 
-def apply(config: Config, actions: dict[str, Action]):
-    config.input
-    config.wem
+def cvt_external(config: Config, actions: HashedActions):
+    from_path = path.join(config.input, config._from["language"])
+    to_path = path.join(config.input, config.to["language"])
+    assert path.isdir(from_path)
+    assert path.isdir(to_path)
+    for ext in listdir(to_path):
+        if not ext.startswith("External"):
+            continue
+        dir = path.join(to_path, ext)
+        for item in listdir(dir):
+            hash, fext = item.split(".")
+            if fext != "wem" or hash not in actions:
+                continue
 
-    rmtree(config.wem)
+            actions[hash].dst_ext = ext
+    return actions
+
+
+def apply(src: str, config: Config, actions: HashedActions):
+    if path.isdir(config.wem):
+        rmtree(config.wem)
     for hash, item in actions.items():
-        makedirs(path.dirname(path.join(config.wem, item.dst_ext)), exist_ok=True)
-        from_path = path.join(config.input, config._from["language"])
+        ext_dir = path.join(config.wem, item.dst_ext)
+        if not path.isdir(ext_dir):
+            makedirs(ext_dir)
         copyfile(
-            path.join(from_path, item.src_ext, item.src_hash + ".wem"),
+            path.join(src, item.src_ext, item.src_hash + ".wem"),
             path.join(config.wem, item.dst_ext, item.dst_hash + ".wem"),
         )
